+++ app/src/main/java/online/partyrun/partyrunapplication/app/ui/PartyRunMain.kt
@@ -23,7 +23,7 @@ fun PartyRunMain(
 ) {
     PartyRunApplicationTheme {
         /**
-             * Navigation Architecture Component
+         * Navigation Architecture Component
          * 메인 기준 NavHostController 인스턴스 생성
          */
         val navController = rememberNavController() // 백 스택 관리 및 현재 목적지가 어떤 Composable인지 추적
