+++ app/src/main/java/online/partyrun/partyrunapplication/core/data/repository/TestRepository.kt
@@ -1,4 +1,4 @@
-package online.partyrun.partyrunapplication.core.domain.repository
+package online.partyrun.partyrunapplication.core.data.repository
 
 import kotlinx.coroutines.flow.Flow
 import online.partyrun.partyrunapplication.core.common.network.ApiResponse
