+++ app/src/main/java/online/partyrun/partyrunapplication/core/data/repository/TestRepositoryImpl.kt
@@ -1,7 +1,6 @@
 package online.partyrun.partyrunapplication.core.data.repository
 
-import online.partyrun.partyrunapplication.core.domain.repository.TestRepository
-import online.partyrun.partyrunapplication.core.network.datasource.TestDataSource
+import online.partyrun.partyrunapplication.core.data.datasource.TestDataSource
 import online.partyrun.partyrunapplication.core.common.network.apiRequestFlow
 import javax.inject.Inject
