+++ app/src/main/java/online/partyrun/partyrunapplication/core/domain/SignInUseCase.kt
@@ -1,9 +1,9 @@
-package online.partyrun.partyrunapplication.core.domain.usecase
+package online.partyrun.partyrunapplication.core.domain
 
 import kotlinx.coroutines.flow.Flow
 import online.partyrun.partyrunapplication.core.model.GoogleIdToken
 import online.partyrun.partyrunapplication.core.model.SignInTokenResponse
-import online.partyrun.partyrunapplication.core.domain.repository.SignInRepository
+import online.partyrun.partyrunapplication.core.data.repository.SignInRepository
 import online.partyrun.partyrunapplication.core.common.network.ApiResponse
 import javax.inject.Inject
